% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gi_calls.R
\name{add_fdrs}
\alias{add_fdrs}
\title{Adds FDR columns to the genetic interaction data}
\usage{
add_fdrs(gi_data, nn_pair_type = "broad")
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{nn_pair_type}{"null distribution" of GI scores used to
compute FDR.Either 'broad' for all interactions with neutral
pairs or 'narrow' for only neutral-neutral pairs}
}
\value{
gi_data with 'FDR.internal' metrics added
}
\description{
Adds FDR columns to the genetic interaction data
}

