% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_gi_calls.R
\name{delta_z_histogram}
\alias{delta_z_histogram}
\title{Histogram of differential interaction effect sizes for neutral and DDR pairs}
\usage{
delta_z_histogram(differential_calls, max_val = 5, min_val = -5,
  nbreaks = 300)
}
\arguments{
\item{differential_calls}{differential interaction dataframe}

\item{max_val}{values above this are grouped in the >= max_val bin}

\item{min_val}{values below this are grouped in the <= min_val bin}

\item{nbreaks}{number of breaks in the histogram}
}
\value{
NULL, makes a plot
}
\description{
Histogram of differential interaction effect sizes for neutral and DDR pairs
}
