% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_gi_calls.R
\name{differential_gi_analysis}
\alias{differential_gi_analysis}
\title{Differential Genetic Interaction analysis}
\usage{
differential_gi_analysis(gi_data, fdr_cutoff = 0.05, delta_gi_cutoff = 0,
  require_sign_change = T, nn_pair_type = "broad", make_plots = F)
}
\arguments{
\item{gi_data}{processed genetic interaction data}

\item{fdr_cutoff}{false discovery rate cutoff for calling significant GI changes}

\item{require_sign_change}{require different GI classifications to call a differential interaction; True or False}

\item{nn_pair_type}{null distribution" of GI scores used to
compute FDR.  Either 'broad' for all interactions with neutral
pairs or 'narrow' for only neutral-neutral pairs}
}
\value{
a data frame with differential genetic interaction comparisons
}
\description{
Differential Genetic Interaction analysis
}

