% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_linked_vs_unlinked.R
\name{gis_vs_linkage_hist}
\alias{gis_vs_linkage_hist}
\title{Make a histogram of GIS vs physical distance}
\usage{
gis_vs_linkage_hist(gi_data, col_same_same = rgb(91/255, 179/255, 228/255,
  0.7), col_linked = rgb(230/255, 154/255, 36/255, 0.7),
  col_unlinked = rgb(0, 153/255, 128/255, 0.7), linkage_cutoff = 75000)
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{col_same_same}{histogram colour for same-same pairs}

\item{col_linked}{histogram colour for linked pairs}

\item{col_unlinked}{histogram colour for unlinked pairs}

\item{linkage_cutoff}{cutoff (bp) for considering two genes on the same chromsome as linked}
}
\value{
NULL, makes a plot
}
\description{
Make a histogram of GIS vs physical distance
}
