% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_onge_comparison_plots.R
\name{prec_rec_vs_stonge}
\alias{prec_rec_vs_stonge}
\title{#' Plots St. Onge et al validation as a function of the internal FDR estimates
#'
#' @param gi_data input genetic interaction table
#' @param control_name condition which corresponds to 'NoDrug' in the St onge data
#' @param condition_name condition which corresponds to 'MMS' in the St onge data
#' @param fdr_prefix 
#' @param z_prefix 
#' @param fdr_cutoff 
#' @param xlims plot x axis limits
#' @param gi_prefix p
#' @param metr performance metric string, must be a valid ROCR metric
#' @param ylab plot y axis limits
#' @param draw_cutoffs boolean, whether to draw lines at cutoffs 
#' @param gi_cutoff_pos 
#' @param gi_cutoff_neg 
#' @param cutoffs_drawn 
#' @param xlab plot x axis label
Plot two ROCR metrics simultaneously as a function of GIS cutoff
(defaults to precision and recall)}
\usage{
prec_rec_vs_stonge(gi_data, control_name = "NoDrug", condition_name = "MMS",
  fdr_prefix = "FDR.neutral_xy", z_prefix = "Z_GIS_xy",
  gi_prefix = "GIS_xy", fdr_cutoff = 0.01, metr1 = "prec",
  metr2 = "rec", metr1_name = "Precision", metr2_name = "Recall",
  col1 = "#C28202", col2 = "#7C32C2", xlims = c(-0.1, 0.1), ylims = c(0,
  100), xlab = "GIS Cutoff", ylab1 = "Precision (\%)",
  ylab2 = "Recall (\%)", draw_cutoffs = T, legend_pos = c(0, 27),
  cutoffs_drawn = c(-0.07, 0.07))
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{control_name}{condition name in gi_data which corresponds to 'NoDrug' in the St onge data}

\item{condition_name}{condition name in gi_data which corresponds to 'MMS' in the St onge data}

\item{fdr_prefix}{pasted with control_name and condition_name to find the FDR columns}

\item{z_prefix}{pasted with control_name and condition_name to find the Z columns}

\item{gi_prefix}{pasted with control_name and condition_name to find the GIS columns}

\item{fdr_cutoff}{strains above this FDR are filtered out when drawing the plot}

\item{metr1}{first metric to be plotted as function of GIS. Must be valid ROCR performance string. defaults to 'prec'}

\item{metr2}{second metric to be plotted as function of GIS. Must be valid ROCR performance string. defaults to 'rec'}

\item{metr1_name}{plot name for first metric}

\item{metr2_name}{plot name for second metric}

\item{col1}{line colour for drawing first metric performance}

\item{col2}{line colour for drawing second metric performance}

\item{xlims}{x limits for plot}

\item{ylims}{y limits for plot}

\item{xlab}{x label for plot}

\item{ylab1}{left y label for plot}

\item{ylab2}{right y label for plot}

\item{draw_cutoffs}{boolean, whether draw lines at GIS cutoffs in cutoffs_drawn}

\item{legend_pos}{where to position the legend}

\item{cutoffs_drawn}{where to draw lines and summarize perforamnce}
}
\value{
NULL, makes a plot
}
\description{
#' Plots St. Onge et al validation as a function of the internal FDR estimates
#'
#' @param gi_data input genetic interaction table
#' @param control_name condition which corresponds to 'NoDrug' in the St onge data
#' @param condition_name condition which corresponds to 'MMS' in the St onge data
#' @param fdr_prefix 
#' @param z_prefix 
#' @param fdr_cutoff 
#' @param xlims plot x axis limits
#' @param gi_prefix p
#' @param metr performance metric string, must be a valid ROCR metric
#' @param ylab plot y axis limits
#' @param draw_cutoffs boolean, whether to draw lines at cutoffs 
#' @param gi_cutoff_pos 
#' @param gi_cutoff_neg 
#' @param cutoffs_drawn 
#' @param xlab plot x axis label
Plot two ROCR metrics simultaneously as a function of GIS cutoff
(defaults to precision and recall)
}

