% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_onge_comparison_plots.R
\name{precision_vs_stonge}
\alias{precision_vs_stonge}
\title{Plots St. Onge et al validation as a function of the internal FDR estimates}
\usage{
precision_vs_stonge(gi_data, control_name = "NoDrug",
  condition_name = "MMS", fdr_prefix = "FDR.Internal_ij",
  z_prefix = "Z_GIS_ij", fdr_cutoff = 0.05, xlims = c(-5, 5),
  xlab = "-Log10 (Internal FDR)*sign(Z)")
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{control_name}{condition which corresponds to 'NoMMS' in the St onge data}

\item{condition_name}{condition which corresponds to 'MMS' in the St onge data}

\item{fdr_prefix}{pasted with control_name and condition_name to find the FDR columns}

\item{z_prefix}{pasted with control_name and condition_name to find the Z columns}

\item{fdr_cutoff}{used to draw 'cutoff' lines in the plot}

\item{xlims}{plot x axis limits}

\item{xlab}{plot x axis label}
}
\description{
Plots St. Onge et al validation as a function of the internal FDR estimates
}

