% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_onge_comparison_plots.R
\name{st_onge_auc_plot}
\alias{st_onge_auc_plot}
\title{Plots FPR-Sensitivty curve versus the St. Onge data}
\usage{
st_onge_auc_plot(gi_data, control_name = "NoDrug", condition_name = "MMS",
  fdr_prefix = "FDR.Internal_ij", z_prefix = "Z_GIS_ij", old_data = F,
  gi_prefix = "GIS_ij", neg_col = rgb(230/255, 155/255, 34/255),
  pos_col = rgb(90/255, 179/255, 228/255), lwd = 3)
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{control_name}{condition which corresponds to 'NoMMS' in the St onge data}

\item{condition_name}{condition which corresponds to 'MMS' in the St onge data}

\item{fdr_prefix}{pasted with control_name and condition_name to find the FDR columns}

\item{z_prefix}{pasted with control_name and condition_name to find the Z columns}

\item{old_data}{if true, uses the raw GI scores, rather than FDR scores to do AUC}

\item{gi_prefix}{pasted with control_name and condition_name to find the raw GI columns}

\item{neg_col}{colour to draw performance curve for negative interactions}

\item{pos_col}{colour to draw performance curve for positive interactions}

\item{lwd}{line width in the plot}
}
\description{
Plots FPR-Sensitivty curve versus the St. Onge data
}

