% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_onge_comparison_plots.R
\name{st_onge_scatterplot}
\alias{st_onge_scatterplot}
\title{Creates a scatterplot of genetic interaction scores vs St. Onge}
\usage{
st_onge_scatterplot(gi_data, control_name = "NoDrug",
  condition_name = "MMS", gi_prefix = "GIS_xy")
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{control_name}{condition which corresponds to 'NoMMS' in the St onge data}

\item{condition_name}{condition which corresponds to 'MMS' in the St onge data}

\item{gi_prefix}{pasted with control_name and condition_name to find the raw GI columns}
}
\description{
Creates a scatterplot of genetic interaction scores vs St. Onge
}

