% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gis.R
\name{update_gis}
\alias{update_gis}
\title{Update BFG-GI genetic interaction data}
\usage{
update_gis(gi_data, pseudocount = 0.5, count_data_grep_pattern = "^C_",
  g_wt_vec)
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{pseudocount}{integer, pseudocount to add in the calculations}

\item{count_data_grep_pattern}{grep pattern used to find count columns}

\item{g_wt_vec}{how many generations each pool grew, in the same order as the count columns}
}
\value{
an updated gi_data
}
\description{
Update BFG-GI genetic interaction data
}

