% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gi_calls.R
\name{add_p_values}
\alias{add_p_values}
\title{Adds p-value columns to the genetic interaction data}
\usage{
add_p_values(gi_data, nn_pair_type = "broad", z_grep_pattern = "^Z_GIS",
  gi_grep_pattern = "^GIS", make_plots = F)
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{nn_pair_type}{"null distribution" of GI scores used to
compute FDR.Either 'broad' for all interactions with neutral
pairs or 'narrow' for only neutral-neutral pairs}
}
\value{
gi_data with 'FDR.internal' metrics added
}
\description{
Adds p-value columns to the genetic interaction data
}

