% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gi_calls.R
\name{update_calls}
\alias{update_calls}
\title{Update genetic interaction calls based on either Z metrics or internal FDR}
\usage{
update_calls(gi_data, z_score_cols = NULL, fdr_cols = NULL,
  gi_cols = NULL, fdr_cutoff_pos = 0.05, fdr_cutoff_neg = 0.05,
  z_cutoff_neg = 0, z_cutoff_pos = 0, gi_cutoff_neg = 0,
  gi_cutoff_pos = 0)
}
\arguments{
\item{gi_data}{input genetic interaction table}

\item{z_score_cols}{which columns correspond to Z scores - default is to use grep to find automatically}

\item{fdr_cols}{which columns correspond to internal FDRs - default is to use grep to find automatically}

\item{z_cutoff_neg}{if use_z is True, Z-cutoff for negative interactions}

\item{z_cutoff_pos}{if use_z is True, Z-cutoff for positive interactions}

\item{z_class_cols}{which columns correspond to Z-based classifications - default is to use grep to find automatically}

\item{fdr_cutoff}{internal FDR cutoff to make GI calls - defaults to 0.05}

\item{use_z}{use Z scores instead of FDR to make GI calls? defaults to False}
}
\value{
a version of gi_data with GI calls updated
}
\description{
Update genetic interaction calls based on either Z metrics or internal FDR
}

