% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gis.R
\name{update_gis}
\alias{update_gis}
\title{Update BFG-GI genetic interaction data}
\usage{
update_gis(gi_data, pseudocount = 1, count_data_grep_pattern = "^C_",
  g_wt_vec = c(12.62, 8.34, 8.44, 7.04, 7.7, 7.84, 7.5, 7.76, 6.94, 6.28))
}
\arguments{
\item{gi_data}{BFG-GI dataframe}

\item{pseudocount}{integer, pseudocount to add in the calculations}

\item{count_data_grep_pattern}{grep pattern used to find count columns}

\item{g_wt_vec}{how many generations each pool grew, in the same order as the count columns}
}
\value{
an updated gi_data
}
\description{
Update BFG-GI genetic interaction data
}

